Refactor: MVC 구조 재정립
