+++ src/js/View/view.js
@@ -1,6 +1,6 @@
-import { CATEGORIES, LOCALSTORAGE } from '../constants/constants.js';
+import { LOCALSTORAGE } from '../constants/constants.js';
 import { $ } from '../common/DOM.js';
-import { getLocalStorage, setLocalStroage } from '../common/localStorage.js';
+import { getLocalStorage } from '../common/localStorage.js';
 import Model from '../Model/model.js';
 import { menuListTemplate } from '../common/template.js';
 
@@ -9,19 +9,18 @@ export default class View {
     this.Model = new Model();
 
     this.$ = {
-      menuInput: $('#menu-name'),
       menuList: $('#menu-list'),
-      menuCount: $('.menu-count'),
     };
   }
 
-  initApp() {
+  clearInputValue(input) {
+    input.value = '';
+  }
+
+  updateMenuCount(target, category) {
     const storage = getLocalStorage(LOCALSTORAGE.ITEM);
 
-    if (storage === null) {
-      setLocalStroage(LOCALSTORAGE.ITEM, this.Model.menu);
-    }
-    this.render(CATEGORIES.ESPRESSO.EN);
+    target.textContent = `총 ${storage[category].length}개`;
   }
 
   render(category) {
@@ -38,8 +37,5 @@ export default class View {
         menuListTemplate(item, count)
       );
     });
-
-    this.Model.clearInputValue(this.$.menuInput);
-    this.Model.updateMenuCount(this.$.menuCount, category);
   }
 }
